+++ src/sqlite3mc_amalgamation.h
@@ -30,10 +30,10 @@
 #define SQLITE3MC_VERSION_H_
 
 #define SQLITE3MC_VERSION_MAJOR      2
-#define SQLITE3MC_VERSION_MINOR      0
-#define SQLITE3MC_VERSION_RELEASE    4
+#define SQLITE3MC_VERSION_MINOR      1
+#define SQLITE3MC_VERSION_RELEASE    0
 #define SQLITE3MC_VERSION_SUBRELEASE 0
-#define SQLITE3MC_VERSION_STRING     "SQLite3 Multiple Ciphers 2.0.4"
+#define SQLITE3MC_VERSION_STRING     "SQLite3 Multiple Ciphers 2.1.0"
 
 #endif /* SQLITE3MC_VERSION_H_ */
 /*** End of #include "sqlite3mc_version.h" ***/